Split up the imagelibrary source file.
The contents are now in two separate files, making the organization a
bit more obvious.
