+++ src/mosaic/_types.py
@@ -1,0 +1,5 @@
+import pathlib
+from typing import Union
+
+
+PathLike = Union[str, pathlib.Path]
