+++ src/mosaic/imagelibrary/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from .library import ImageLibrary
+from .cifar100library import CIFAR100Library
